Fix incorrect child project's ppf file path stored in paf file.
While generating the .paf file, the child project's ppf file shall be constructed and stored. But this ppf file path is generated with the application's root folder instead of the project's root folder.
Note, the ppf file itself gets generated in the right location, but when that information is stored in the paf file it is incorrect.

I noticed that the OunceProjectBean was feeded with the application's path instead of the project's path, which was causing an incorrect ppf file path generation.